Refactor: rename integration test class and move to correct package #3007
