+++ nexus-ingestion-api/src/test/java/org/techbd/ingest/integrationtests/controller/XdsRepositoryControllerITCase.java
@@ -1,4 +1,4 @@
-package org.techbd.ingest.integrationtests.soap;
+package org.techbd.ingest.integrationtests.controller;
 
 
 import java.nio.charset.StandardCharsets;
